+++ server/api/schema.py
@@ -1,0 +1,5 @@
+import graphene
+from api.queries.queries import Query
+from api.mutations.mutation import Mutation
+
+schema = graphene.Schema(query=Query, mutation=Mutation)
